requirements.txt
dj-database-url==0.5.0
Django==1.9
django-heroku==0.3.1
djangorestframework==3.6.3
numpy==1.18.1
pandas==1.0.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
whitenoise==5.0
